udpate README and refactor
