+++ run_DQ.sh
@@ -6,4 +6,4 @@ run=$1
 output_dir=$2
 
 cp /home/ppd/bewilson/FASER_2024_DQ_Analyis/RDFDefines.h .
-python3 /home/ppd/bewilson/FASER_2024_DQ_Analyis/FASER_DQ_RDF_improved.py $run -o $output_dir
+python3 /home/ppd/bewilson/FASER_2024_DQ_Analyis/FASER_DQ_RDF.py $run -o $output_dir
