--- sample_lists/2022_r0017.yaml
@@ -1,105 +1,0 @@
-samples:
-  - /home/faser/data/phys/2022_unb/r0017/008095/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008115/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008119/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008120/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008124/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008134/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008137/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008152/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008153/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008155/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008157/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008174/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008175/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008179/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008182/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008187/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008191/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008193/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008251/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008252/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008266/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008269/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008290/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008291/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008294/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008297/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008301/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008320/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008326/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008327/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008328/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008330/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008331/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008332/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008333/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008334/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008336/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008715/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008717/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008725/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008728/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008729/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008730/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008731/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008733/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008736/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008749/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008752/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008775/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008777/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008784/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008798/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008799/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008804/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008834/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008837/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008841/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008843/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008846/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008847/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008848/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008850/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008906/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008908/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008912/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008913/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008915/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008916/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008917/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008930/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008933/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008934/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008943/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008944/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008945/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008947/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008948/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008949/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008950/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008952/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008955/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008969/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008970/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008972/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008975/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008977/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008981/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008985/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/008987/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009058/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009060/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009063/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009069/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009072/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009073/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009092/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009094/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009148/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009163/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009166/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009176/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009181/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009183/Faser-Physics-*-PHYS.root
-  - /home/faser/data/phys/2022_unb/r0017/009188/Faser-Physics-*-PHYS.root
