--- sample_lists/2023_r0017.yaml
@@ -1,101 +1,0 @@
-samples:
-  # - /home/faser/data/phys/2023/r0017/010417/Faser-Physics-010417-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010419/Faser-Physics-010419-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010422/Faser-Physics-010422-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010424/Faser-Physics-010424-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010426/Faser-Physics-010426-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010443/Faser-Physics-010443-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010526/Faser-Physics-010526-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010538/Faser-Physics-010538-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010540/Faser-Physics-010540-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010555/Faser-Physics-010555-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010572/Faser-Physics-010572-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010600/Faser-Physics-010600-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010601/Faser-Physics-010601-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010602/Faser-Physics-010602-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010604/Faser-Physics-010604-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010605/Faser-Physics-010605-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010615/Faser-Physics-010615-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010693/Faser-Physics-010693-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010694/Faser-Physics-010694-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010695/Faser-Physics-010695-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010698/Faser-Physics-010698-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010699/Faser-Physics-010699-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010700/Faser-Physics-010700-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010701/Faser-Physics-010701-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010729/Faser-Physics-010729-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010732/Faser-Physics-010732-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010733/Faser-Physics-010733-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010738/Faser-Physics-010738-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010743/Faser-Physics-010743-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010747/Faser-Physics-010747-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010775/Faser-Physics-010775-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010792/Faser-Physics-010792-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010794/Faser-Physics-010794-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010796/Faser-Physics-010796-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/010799/Faser-Physics-010799-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010829/Faser-Physics-010829-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010847/Faser-Physics-010847-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010849/Faser-Physics-010849-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010857/Faser-Physics-010857-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010859/Faser-Physics-010859-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010881/Faser-Physics-010881-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010884/Faser-Physics-010884-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010888/Faser-Physics-010888-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010921/Faser-Physics-010921-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010928/Faser-Physics-010928-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010955/Faser-Physics-010955-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/010956/Faser-Physics-010956-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011038/Faser-Physics-011038-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011041/Faser-Physics-011041-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011068/Faser-Physics-011068-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011070/Faser-Physics-011070-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011079/Faser-Physics-011079-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011081/Faser-Physics-011081-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011082/Faser-Physics-011082-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011086/Faser-Physics-011086-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011088/Faser-Physics-011088-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011089/Faser-Physics-011089-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011093/Faser-Physics-011093-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011096/Faser-Physics-011096-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011098/Faser-Physics-011098-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011127/Faser-Physics-011127-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011156/Faser-Physics-011156-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011208/Faser-Physics-011208-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011211/Faser-Physics-011211-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011212/Faser-Physics-011212-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011213/Faser-Physics-011213-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011214/Faser-Physics-011214-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011243/Faser-Physics-011243-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011461/Faser-Physics-011461-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011463/Faser-Physics-011463-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011471/Faser-Physics-011471-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011478/Faser-Physics-011478-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011480/Faser-Physics-011480-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011486/Faser-Physics-011486-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011488/Faser-Physics-011488-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011491/Faser-Physics-011491-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011504/Faser-Physics-011504-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011506/Faser-Physics-011506-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011519/Faser-Physics-011519-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011523/Faser-Physics-011523-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011527/Faser-Physics-011527-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011528/Faser-Physics-011528-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011583/Faser-Physics-011583-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011584/Faser-Physics-011584-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011585/Faser-Physics-011585-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011586/Faser-Physics-011586-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011589/Faser-Physics-011589-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011590/Faser-Physics-011590-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011595/Faser-Physics-011595-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011596/Faser-Physics-011596-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011597/Faser-Physics-011597-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011599/Faser-Physics-011599-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011602/Faser-Physics-011602-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011613/Faser-Physics-011613-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011630/Faser-Physics-011630-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011700/Faser-Physics-011700-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011701/Faser-Physics-011701-*PHYS.root
-  - /home/faser/data/phys/2023/r0017/011703/Faser-Physics-011703-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011705/Faser-Physics-011705-*PHYS.root
-  # - /home/faser/data/phys/2023/r0017/011706/Faser-Physics-011706-*PHYS.root
